+++ src/config/provider.config.ts
@@ -1,0 +1,7 @@
+// free rpc provider for local development
+export enum Provider {
+	Cloudflare = 'https://cloudflare-eth.com/',
+	Binance = 'https://bsc-dataseed.binance.org/',
+	Quicknode = 'https://rpc-mainnet.matic.quiknode.pro',
+	xDai = 'https://rpc.xdaichain.com',
+}