+++ src/mobx/model/charts/market-delta.ts
@@ -1,0 +1,5 @@
+export interface MarketDelta {
+	date: Date;
+	close: number;
+	change: number;
+}