+++ src/pages/Landing.tsx
@@ -2,7 +2,7 @@ import CurrencyInfoCard from '../components-v2/common/CurrencyInfoCard';
 import CurrencyPicker from '../components-v2/landing/CurrencyPicker';
 import SamplePicker from '../components-v2/landing/SamplePicker';
 import WalletSlider from '../components-v2/landing/WalletSlider';
-import { Grid, Container, makeStyles, Button } from '@material-ui/core';
+import { Grid, makeStyles, Button } from '@material-ui/core';
 import PageHeader from '../components-v2/common/PageHeader';
 import { StoreContext } from '../mobx/store-context';
 import { observer } from 'mobx-react-lite';
@@ -11,17 +11,9 @@ import BigNumber from 'bignumber.js';
 import SettList from 'components-v2/landing/SettList';
 import { RewardsModal } from '../components-v2/landing/RewardsModal';
 import { SettState } from '../mobx/model/setts/sett-state';
+import { HeaderContainer, LayoutContainer } from '../components-v2/common/Containers';
 
 const useStyles = makeStyles((theme) => ({
-	landingContainer: {
-		[theme.breakpoints.up('md')]: {
-			paddingLeft: theme.spacing(30),
-		},
-	},
-	headerContainer: {
-		marginTop: theme.spacing(3),
-		marginBottom: theme.spacing(3),
-	},
 	marginTop: {
 		marginTop: theme.spacing(3),
 	},
@@ -88,8 +80,9 @@ const Landing = observer((props: LandingProps) => {
 	const { title, subtitle, state } = props;
 
 	const {
-		wallet: { connectedAddress, network },
+		wallet: { connectedAddress },
 		uiState: { currency },
+		network: { network },
 		setts,
 		prices,
 		user,
@@ -97,17 +90,18 @@ const Landing = observer((props: LandingProps) => {
 	const { protocolSummary } = setts;
 	const userConnected = !!connectedAddress;
 
+	const badgerToken = network.deploy.token.length > 0 ? network.deploy.token : undefined;
 	const totalValueLocked = protocolSummary ? new BigNumber(protocolSummary.totalValue) : undefined;
-	const badgerPrice = network.deploy ? prices.getPrice(network.deploy.token) : undefined;
+	const badgerPrice = badgerToken ? prices.getPrice(badgerToken) : undefined;
 	const portfolioValue = userConnected && user.initialized ? user.portfolioValue : undefined;
 
 	return (
-		<Container className={classes.landingContainer}>
+		<LayoutContainer>
 			{/* Landing Metrics Cards */}
 			<Grid container spacing={1} justify="center">
-				<Grid item xs={12} className={classes.headerContainer}>
+				<HeaderContainer item xs={12}>
 					<PageHeader title={title} subtitle={subtitle} />
-				</Grid>
+				</HeaderContainer>
 				<Grid item xs={12} className={classes.widgetContainer}>
 					<div className={classes.walletContainer}>{userConnected && <WalletSlider />}</div>
 					<div className={classes.pickerContainer}>
@@ -138,7 +132,7 @@ const Landing = observer((props: LandingProps) => {
 			)}
 
 			<SettList state={state} />
-		</Container>
+		</LayoutContainer>
 	);
 });