+++ src/tests/PageHeader.test.tsx
@@ -1,17 +1,10 @@
 import React from 'react';
 import PageHeader from '../components-v2/common/PageHeader';
-import store from '../mobx/store';
 import '@testing-library/jest-dom';
-import { StoreProvider } from '../mobx/store-context';
-import { customRender } from './Utils';
+import { checkSnapshot } from './utils/snapshots';
 
 test('Renders correctly', () => {
 	const title = 'Test title';
 	const subtitle = 'Test subtitle';
-	const { container } = customRender(
-		<StoreProvider value={store}>
-			<PageHeader title={title} subtitle={subtitle} />
-		</StoreProvider>,
-	);
-	expect(container).toMatchSnapshot();
+	checkSnapshot(<PageHeader title={title} subtitle={subtitle} />);
 });