+++ src/tests/utils/snapshots.tsx
@@ -1,6 +1,6 @@
 import { customRender } from 'tests/Utils';
 import { StoreProvider } from 'mobx/store-context';
-import store from 'mobx/store';
+import store from 'mobx/RootStore';
 import React from 'react';
 
 export const TEST_ADDRESS = '0x03f7724180AA6b939894B5Ca4314783B0b36b329';